handle inline-SVGs (no object tag)
this is handy for Moodle pasting... but need to be
careful: if the KeyshapeJS code is inside the SVG
there's a collision if there are duplicates. For
now, that's catastrophic so the offending diagrams
are removed (!) but the first one is probably
broken too. So it's all bad. Maybe the last one
is the one to try to save, but we're discovering
them in the wrong order.

Thanks for KeyshapeJS's own dev support for some
guidance to get this far!